+++ BlamSauce/game/engines/engine_oddball.hpp
@@ -15,12 +15,12 @@ namespace Yelo
 	{
 		struct s_oddball_globals
 		{
-			int32 score_to_win;
-			int32 team_score[Enums::k_multiplayer_maximum_players];
-			int32 individual_score[Enums::k_multiplayer_maximum_players];
-			int32 undocumented1[Enums::k_multiplayer_maximum_players];
+			long score_to_win;
+			long team_score[Enums::k_multiplayer_maximum_players];
+			long individual_score[Enums::k_multiplayer_maximum_players];
+			long undocumented1[Enums::k_multiplayer_maximum_players];
 			datum_index current_ball_owners[Enums::k_multiplayer_maximum_players]; // player_index
-			int32 current_ball_held_times[Enums::k_multiplayer_maximum_players]; // game-ticks
+			long current_ball_held_times[Enums::k_multiplayer_maximum_players]; // game-ticks
 		}; //static_assert( sizeof(s_slayer_globals) == 0x );
 	};
 };
