+++ BlamSauce/game/game.hpp
@@ -24,10 +24,10 @@ namespace Yelo {
 			// 0x0 long_enum game_mode; // where _game_mode_campaign = 1 (and default). k_number_of_game_modes = 6
 			// 0x4 short game_simulation; // _game_simulation_local is default
 
-			UNKNOWN_TYPE(int32); // never see this referenced or explicitly set. due to memset, this will always be set to zero
-			UNKNOWN_TYPE(int16); // never see this referenced besides in game_options_new code where it's set to 0, even after a memset call. highly likely to be an enum field, albeit unused
+			UNKNOWN_TYPE(long); // never see this referenced or explicitly set. due to memset, this will always be set to zero
+			UNKNOWN_TYPE(short); // never see this referenced besides in game_options_new code where it's set to 0, even after a memset call. highly likely to be an enum field, albeit unused
 			Enums::game_difficulty_level difficulty_level;
-			int32                        game_random_seed;
+			long                        game_random_seed;
 			// not always the actual path, sometimes just the name, go figure
 			char                         map_name[/*Enums::k_max_tag_name_length*/255 + 1];
 		}; static_assert(sizeof(s_game_options) == 0x10C);
